Fix TCP keepalive handling on Haiku and OpenBSD
On Haiku and OpenBSD don't call `setsockopt(fd, IPPROTO_TCP,
SO_KEEPALIVE, secs)`, it is incorrect because `SO_KEEPALIVE` belongs
to `SOL_SOCKET` layer, because it accepts a boolean instead of the
number of seconds, and because there is no way to set keepalive idle
time per socket on these operating systems.

On Haiku and OpenBSD it is only possible to enable and disable
keepalives, and it is already done with `setsockopt(fd, SOL_SOCKET,
SO_KEEPALIVE, 1)` in a portable way in `Socket.set_keepalive()`
defined in `src/socket.rs`.